Temporary fix of usage of state param during auth
In commit b538beaea92dd1dd0bf6b0a7ff229d1b234db8c3, we used the state
param in the URL hash to redirect to schedule pages, so that users that
logged in from the schedule page would be redirected to that schedule page upon
successful login.

However, now that we are receiving a token directly from AD and not from
auth0, we are always getting a state in the hash, and we shouldn't be
redirecting there every time.

We should pass a more descriptive state in order to redirect to a
schedule page, and also use the state to verify the login request.